+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 * Add `EveOnline::ESI::UniverseGraphic`
 * Add `EveOnline::ESI::UniverseSystem`
 * Add `EveOnline::ESI::Models::System`
-* Add `EveOnline::ESI::Models::Planet`
+* Add `EveOnline::ESI::Models::PlanetShort`
 * Add `EveOnline::ESI::Models::Planets`
 
 ## v0.19.0